requirements.txt
beautifulsoup4==4.10.0
Pillow==9.0.0
requests==2.26.0
wordcloud==1.8.1
